requirements.txt
streamlit
pandas
opencv-python
pygame
pyyaml
requests
